Add async APIs to internal IDirectory and IFile
Also move Wrappers folder into Internal to match namespaces
