+++ DNN Platform/Library/DotNetNuke.Library.csproj
@@ -817,6 +817,12 @@
     <Compile Include="Services\FileSystem\Internal\FileSecurityController.cs" />
     <Compile Include="Services\FileSystem\Internal\SecurityCheckers\SvgFileChecker.cs" />
     <Compile Include="Services\FileSystem\Internal\UserSecurityController.cs" />
+    <Compile Include="Services\FileSystem\Internal\Wrappers\DirectoryWrapper.cs" />
+    <Compile Include="Services\FileSystem\Internal\Wrappers\FileWrapper.cs" />
+    <Compile Include="Services\FileSystem\Internal\Wrappers\IconControllerWrapper.cs" />
+    <Compile Include="Services\FileSystem\Internal\Wrappers\IDirectory.cs" />
+    <Compile Include="Services\FileSystem\Internal\Wrappers\IFile.cs" />
+    <Compile Include="Services\FileSystem\Internal\Wrappers\IIconController.cs" />
     <Compile Include="Services\FileSystem\LinkClickPortalSettings.cs" />
     <Compile Include="Services\FileSystem\SyncFolderData.cs" />
     <Compile Include="Services\GeneratedImage\IImageStore.cs" />
@@ -1295,8 +1301,6 @@
     <Compile Include="Services\FileSystem\Exceptions\FolderAlreadyExistsException.cs" />
     <Compile Include="Services\FileSystem\Exceptions\NoNetworkAvailableException.cs" />
     <Compile Include="Services\FileSystem\Internal\DefaultFolderProviders.cs" />
-    <Compile Include="Services\FileSystem\Wrappers\IconControllerWrapper.cs" />
-    <Compile Include="Services\FileSystem\Wrappers\IIconController.cs" />
     <Compile Include="Services\ClientCapability\ClientCapabilityProvider.cs" />
     <Compile Include="Services\ClientCapability\IClientCapability.cs" />
     <Compile Include="Services\Mobile\IMatchRule.cs" />
@@ -1338,11 +1342,7 @@
     <Compile Include="Services\FileSystem\IFileManager.cs" />
     <Compile Include="Services\FileSystem\IFolderInfo.cs" />
     <Compile Include="Services\FileSystem\IFolderManager.cs" />
-    <Compile Include="Services\FileSystem\Wrappers\DirectoryWrapper.cs" />
-    <Compile Include="Services\FileSystem\Wrappers\FileWrapper.cs" />
     <Compile Include="Common\Utilities\ICBO.cs" />
-    <Compile Include="Services\FileSystem\Wrappers\IDirectory.cs" />
-    <Compile Include="Services\FileSystem\Wrappers\IFile.cs" />
     <Compile Include="Services\FileSystem\Providers\DatabaseFolderProvider.cs" />
     <Compile Include="Services\FileSystem\Exceptions\FileAlreadyExistsException.cs" />
     <Compile Include="Services\FileSystem\FileInfo.cs" />
