+++ DNN Platform/Tests/DotNetNuke.Tests.Core/Providers/Folder/DatabaseFolderProviderTests.cs
@@ -14,6 +14,7 @@ namespace DotNetNuke.Tests.Core.Providers.Folder
     using DotNetNuke.Data;
     using DotNetNuke.Entities.Users;
     using DotNetNuke.Services.FileSystem;
+    using DotNetNuke.Services.FileSystem.Internal;
     using DotNetNuke.Tests.Utilities;
     using DotNetNuke.Tests.Utilities.Mocks;
     using Moq;
